requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.8.2
cachetools==5.4.0
cattrs==23.2.3
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
chroma-hnswlib==0.7.3
chromadb==0.4.13
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
dataclasses-json==0.6.1
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
duckduckgo_search==6.2.9
emoji==2.8.0
exceptiongroup==1.1.3
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.103.2
filelock==3.12.4
filetype==1.2.0
flatbuffers==24.3.25
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.11
GitPython==3.1.43
google-search-results==2.4.2
gpt4all==2.0.0
graphlib-backport==1.0.3
greenlet==2.0.2
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.0
httpx==0.27.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-resources==6.1.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.2
jiter==0.5.0
joblib==1.3.2
jq==1.8.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.29
langchain-groq==0.1.9
langchain-openai==0.1.21
langchain-text-splitters==0.2.2
langchainhub==0.1.13
langdetect==1.0.9
langserve==0.2.2
langsmith==0.1.98
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.7
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.6
numpy==1.24.4
onnxruntime==1.16.0
openai==1.40.3
openmeteo_requests==1.3.0
openmeteo_sdk==1.14.1
orjson==3.10.7
overrides==7.4.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
Pillow==10.0.1
platformdirs==4.2.2
posthog==3.0.2
primp==0.6.0
prompt_toolkit==3.0.47
protobuf==4.24.3
psutil==6.0.0
psycopg2==2.9.9
ptyprocess==0.7.0
pulsar-client==3.3.0
pure_eval==0.2.3
pyarrow==17.0.0
pydantic==1.10.13
pydantic_core==2.10.1
pydeck==0.9.1
pydevd==2.10.0
Pygments==2.18.0
pypdf==4.3.1
PyPika==0.48.9
pyproject-toml==0.0.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.1.0
referencing==0.35.1
regex==2023.8.8
requests==2.31.0
requests-cache==1.2.1
retry-requests==2.0.0
rich==13.7.1
rpds-py==0.20.0
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.27.0
streamlit==1.37.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
terminal==0.4.0
threadpoolctl==3.2.0
tiktoken==0.7.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.1
torch==2.0.1
torchvision==0.15.2
tornado==6.4.1
tqdm==4.66.1
traitlets==5.14.3
transformers==4.33.3
typer==0.9.0
types-requests==2.31.0.10
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.10.18
url-normalize==1.4.3
urllib3==2.0.5
uvicorn==0.23.2
uvloop==0.17.0
watchdog==4.0.2
watchfiles==0.20.0
wcwidth==0.2.13
websockets==11.0.3
wikipedia==1.4.0
yarl==1.9.2
zipp==3.17.0
